requirements.txt
Django==3.0.5
drf-yasg==1.17.1
gunicorn==20.0.4
djangorestframework==3.11.0
psycopg2-binary==2.8.5
supervisor==4.2.0
boto3==1.13.16
django-storages==1.9.1
Pillow==7.1.2
twilio==6.40.0
django-multiselectfield==0.1.12
openpyxl==3.0.3
